removed routes, controller and service related to openAI. adjusted documentation codes as well
